Stabilized markers styles
Previous version placed markers on a separate layer and calculated their
positions independently. Now the markers are placed on the same layer as the
tiles and their positions are now relative to that layer so they can be static
and the markers can be translated with the layer as a whole.